AutomatedGrader/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
pdfminer.six
duckduckgo-search
google-genai
python-dotenv
requests